工作流库0216/T06-城投区域分析/requirements.txt
# 城投区域分析 - Python依赖列表

# 数据处理
pandas>=1.5.0
numpy>=1.21.0

# 数据库
sqlalchemy>=1.4.0
pymysql>=1.0.0

# 数据可视化
matplotlib>=3.5.0
seaborn>=0.12.0

# 金融数据处理
openpyxl>=3.0.0

# HTTP请求
requests>=2.28.0

# 日期时间处理
python-dateutil>=2.8.0

# 环境变量管理
python-dotenv>=0.21.0
